Dropped all tut.by filters
Lukashenko's regime killed tut.by which was one of the most important mass media in Belarus. So all the filters related to it as it will not back.
Wondering why they killed it? Because Lukashenko loose elections 2020 but he doesn't want to leave and is trying to hold just by using [ultimate violence](https://august2020.info/)
Nice try, BUT he will be overthrown anyway and Belarus will be free.
Long live Belarus!